third_party/blink/web_tests/external/wpt/selection/bidi/move-by-character-expected.txt

This is a testharness.js-based test.
[FAIL] RTL text in LTR paragraph: move right 2 chars
  assert_equals: expected 2 but got 6
[FAIL] RTL text in LTR paragraph: move left 2 chars
  assert_equals: expected 4 but got 0
[FAIL] LTR text in RTL paragraph: move right 2 chars
  assert_equals: expected 5 but got 1
[FAIL] LTR text in RTL paragraph: move left 2 chars
  assert_equals: expected 3 but got 7
[FAIL] LTR-RTL context in LTR paragraph: move right 3 chars
  assert_equals: expected 11 but got 7
[FAIL] LTR-RTL context in LTR paragraph: move left 3 chars
  assert_equals: expected 4 but got 8
[FAIL] RTL-LTR context in LTR paragraph: move right 3 chars
  assert_equals: expected 14 but got 10
[FAIL] RTL-LTR context in LTR paragraph: move left 3 chars
  assert_equals: expected 7 but got 11
[FAIL] Mixed context in auto-dir paragraph: move right 3 chars
  assert_equals: expected 5 but got 14
[FAIL] Mixed context in auto-dir paragraph: move left 3 chars
  assert_equals: expected 17 but got 8
Harness: the test ran to completion.
